Visualizations-of-Probability-Distributions-master/requirements.txt
streamlit
numpy
matplotlib.pyplot
tensorflow
tensorflow_probability
math